pyproject.toml
[project]
name = "curatedmetagenomicdataetl"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "duckdb>=1.4.1",
    "google-cloud-bigquery>=3.38.0",
    "google-cloud-storage>=2.18.2",
    "httpx>=0.27.0",
    "tqdm>=4.66.0",
]
